refactor datamapper sqlite3 test helper to inherit from DatabaseHelper.
