+++ spec/database_cleaner/data_mapper/truncation/sqlite3_spec.rb
@@ -1,38 +1,30 @@
-require 'dm-core'
-require 'dm-sqlite-adapter'
-require 'support/data_mapper/sqlite3_setup'
+require 'support/data_mapper/sqlite3_helper'
 require 'database_cleaner/data_mapper/truncation'
 
 RSpec.describe DatabaseCleaner::DataMapper::Truncation do
   let(:helper) { DataMapperSQLite3Helper.new }
 
-  let(:connection) do
-    helper.data_mapper_sqlite3_connection
-  end
+  let(:connection) { helper.connection }
 
   around do |example|
-    helper.data_mapper_sqlite3_setup
-
+    helper.setup
     example.run
-
-    helper.data_mapper_sqlite3_teardown
+    helper.teardown
   end
 
   describe "DM connection adapter monkeypatches" do
+    before do
+      2.times { DmUser.create }
+    end
+
     describe "#truncate_table" do
       it "truncates the table" do
-        2.times { DmUser.create }
-
         connection.truncate_table(DmUser.storage_names[:default])
         expect(DmUser.count).to eq 0
       end
 
       it "resets AUTO_INCREMENT index of table" do
-        2.times { DmUser.create }
-        DmUser.destroy
-
         connection.truncate_table(DmUser.storage_names[:default])
-
         expect(DmUser.create.id).to eq 1
       end
     end
